Some changes to how the influx client handles errors.
We now set the timeout for an influx request at 65 sec instead of
blocking forever & waiting. Any passed-in config value is respected
here instead of blown away, but if you pass nothing we set the timeout.

On `writePoint()` we fire the callback on the next tick instead of
waiting for the influx call to return.  Errors from influx are logged,
but not returned to the streaming layer. We wish to attempt to send
points but not stress about failures.

Added tests for the new behavior.