fix: clean up influxdb data on the way out
This should prevent the collector from poisoning batches quite as often
as it does right now. We achieve this by cleaning up tag names on their
way through, and by throwing out tags with null/undefined data.

We need to pin to influx@4.2.2 for the client library because we're
using influx 0.13 in production right now.

Added tests to verify that we're sanitizing.